Run watch command test in parallel
