+++ tests/integration/cmd_watch_test.go
@@ -1,0 +1,51 @@
+package integration
+
+import (
+	"os"
+	"path/filepath"
+	"time"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"github.com/openshift/odo/tests/helper"
+)
+
+var _ = Describe("odoWatchE2e", func() {
+	var project string
+	var context string
+
+	// Setup up state for each test spec
+	// create new project (not set as active) and new context directory for each test spec
+	// This is before every spec (It)
+	BeforeEach(func() {
+		SetDefaultEventuallyTimeout(10 * time.Minute)
+		project = helper.CreateRandProject()
+		context = helper.CreateNewContext()
+		os.Setenv("GLOBALODOCONFIG", filepath.Join(context, "config.yaml"))
+	})
+
+	// Clean up after the test
+	// This is run after every Spec (It)
+	AfterEach(func() {
+		helper.DeleteProject(project)
+		helper.DeleteDir(context)
+		os.Unsetenv("GLOBALODOCONFIG")
+	})
+
+	Context("when running help for watch command", func() {
+		It("should display the help", func() {
+			appHelp := helper.CmdShouldPass("odo", "watch", "-h")
+			Expect(appHelp).To(ContainSubstring("Watch for changes"))
+		})
+	})
+
+	Context("when executing watch without pushing the component", func() {
+		It("should fail", func() {
+			helper.CopyExample(filepath.Join("source", "nodejs"), context)
+			helper.CmdShouldPass("odo", "component", "create", "nodejs", "--project", project, "--context", context)
+			output := helper.CmdShouldFail("odo", "watch", "--context", context)
+			Expect(output).To(ContainSubstring("component does not exist. Please use `odo push` to create you component"))
+		})
+	})
+})
